feat(notifier): implement user, editor, and admin notification system (#460)
* feat(notifier): implement user, editor, and admin notification system

* feat(notifier): enhance notification system with new folder events and email sending functionality

* feat(notifier): replace console.error with logger for email error handling and connection verification

* feat(notifier): simplify sendMessage function by destructuring parameters and improving email handling

* feat(notifier): add notification template and enhance notification sending functionality

* feat(notifier): improve notification messages and rename sendMessage to sendMail for clarity

* feat(notifier): consolidate notification types into a single object for improved structure

* feat(notifier): refactor sendMail function to use destructured parameters and update notification template handling

* feat(notifier): restructure notifier module and add client notification options

* feat(notifier): add sendAdminNotification and formatNotificationOptions functions for enhanced notification handling

* refactor(notifier): remove unused error class and clean up import statements

* chore(resources): remove unused logo and favicon files to clean up assets

* feat(notifier): add user and editor notification functions for account and folder updates

* feat(notifier): add user notification settings component and API for updating notifications

* feat(notifier): prefix notification toggle labels with user roles for clarity

* feat(notifier): implement basic email notifier system for User and Admin notifications

* Apply suggestions from code review


* fix(notifications): correct syntax for verifying user permission levels

---------

+++ packages/studiocms/src/routes/dashboard/studiocms_api/dashboard/reset-password.ts
@@ -1,6 +1,7 @@
 import { verifyPasswordStrength } from 'studiocms:auth/lib/password';
 import { developerConfig } from 'studiocms:config';
 import { apiResponseLogger } from 'studiocms:logger';
+import { sendAdminNotification, sendUserNotification } from 'studiocms:notifier';
 import studioCMS_SDK from 'studiocms:sdk';
 import type { APIContext, APIRoute } from 'astro';
 
@@ -51,10 +52,19 @@ export const POST: APIRoute = async (ctx: APIContext) => {
 		password: password,
 	};
 
+	const userData = await studioCMS_SDK.GET.databaseEntry.users.byId(userid);
+
+	if (!userData) {
+		return apiResponseLogger(404, 'User not found');
+	}
+
 	try {
 		await studioCMS_SDK.AUTH.user.update(userid, userUpdate);
 
 		await studioCMS_SDK.resetTokenBucket.delete(userid);
+
+		await sendUserNotification('account_updated', userid);
+		await sendAdminNotification('user_updated', userData.username);
 
 		return apiResponseLogger(200, 'User password updated successfully');
 	} catch (error) {
